sdk/nscomposer: split endpoint-server
